requirements.txt
Flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
openai==1.3.7
python-dotenv==1.0.0
sentence-transformers==2.6.1
huggingface_hub==0.19.4
transformers==4.36.2
torch==2.0.1
urllib3==1.26.18
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
gunicorn==21.2.0
# macOS Apple Silicon optimized
faiss-cpu==1.7.4; platform_machine != "arm64"
faiss-cpu==1.8.0; platform_machine == "arm64"